backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]